+++ src/control_plane/shared_infrastructure/mongodb/mod.rs
@@ -1,9 +1,15 @@
 pub(crate) use mongodb_logical_resource_plan::MongoDbLogicalResourcePlan;
 pub(crate) use mongodb_plan_error::MongoDbPlanError;
+pub(crate) use mongodb_project_resources::MongoDbProjectResources;
 pub(crate) use mongodb_shared_instance_plan::MongoDbSharedInstancePlan;
 pub(crate) use mongodb_shared_instance_plan_options::MongoDbSharedInstancePlanOptions;
+pub(crate) use plan_mongodb_project_resources::plan_mongodb_project_resources;
+pub(crate) use provision_mongodb_logical_resource::provision_mongodb_logical_resource;
 
 mod mongodb_logical_resource_plan;
 mod mongodb_plan_error;
+mod mongodb_project_resources;
 mod mongodb_shared_instance_plan;
 mod mongodb_shared_instance_plan_options;
+mod plan_mongodb_project_resources;
+mod provision_mongodb_logical_resource;